requirements.txt
accelerate==0.20.3
datasets==2.13.0
decord==0.6.0
einops==0.6.1
evaluate==0.4.0
ftfy==6.1.1
iopath==0.1.10
opencv-python==4.7.0.72
peft @ git+https://github.com/huggingface/peft@08cb3dde577747f6ca6638c884fd66fd16cf2e9d
pytorchvideo @ git+https://github.com/facebookresearch/pytorchvideo.git@28fe037d212663c6a24f373b94cc5d478c8c1a1d
scipy==1.10.1
tensorboardX==2.6.1
tokenizers==0.13.3
tqdm==4.65.0
numpy==1.26.4
scikit-learn==1.3.0
braceexpand==0.1.7
webdataset==0.2.48
transformers==4.30.2
urllib3==1.26.15
soundfile==0.12.1
faiss-gpu==1.7.2
fastparquet==2024.11.0
ranx==0.3.20